Move template handling logic into a service class
Almost done with the refactoring. This was the last big blob of code
that needed to be moved into its own service class. This takes all the
template handling logic and moves it out into a self-contained unit that
deals with retrieving template contents.
